test(querier): move seed fixtures to fixtures.queriercommon and inline assertions
Review feedback: the seed fixtures and the data constants move to
fixtures/queriercommon.py, named after the test package they serve. The
query execution, extraction, and assertion now sit inside each test
function, so a reader sees what each case asserts without following a
helper.
